Move ICA-related files.
